Minor refactoring - Extract presets
- As the preset logics grows, I wanted to extract it from the ddprof_context init
- Minor renaming / namespacing refactoring
