+++ include/ddprof_context_lib.hpp
@@ -10,8 +10,10 @@ typedef struct DDProfInput DDProfInput;
 typedef struct DDProfContext DDProfContext;
 typedef struct PerfWatcher PerfWatcher;
 
+namespace ddprof {
 /***************************** Context Management *****************************/
-DDRes ddprof_context_set(DDProfInput *input, DDProfContext *);
-void ddprof_context_free(DDProfContext *);
+DDRes context_set(DDProfInput *input, DDProfContext *);
+void context_free(DDProfContext *);
 
-int ddprof_context_allocation_profiling_watcher_idx(const DDProfContext *ctx);
+int context_allocation_profiling_watcher_idx(const DDProfContext *ctx);
+} // namespace ddprof
